refactor: split telemetry pipeline into generator, hub, and consumer packages
- Replace client/server into focused generator, hub, and consumer packages.
- Split sensor/processor logic under generator, sender under hub, and listener/parser/logger under consumer.
- Rename the sender file/function for clarity (sender.go -> senderUDP.go, Send -> SendUDP).
- Wire main.go to create the shared result channel and start each stage via Run function.
