+++ internal/consumer/listener.go
@@ -1,4 +1,4 @@
-package client
+package consumer
 
 import (
 	"fmt"
