+++ internal/generator/processor.go
@@ -1,4 +1,4 @@
-package server
+package generator
 
 import (
 	"time"
